+++ varnish-macros/src/metrics.rs
