+++ varnish/src/lib.rs
@@ -101,10 +101,12 @@ pub use varnish_sys::ffi;
 pub mod varnishtest;
 
 mod metrics_reader;
-mod vsc_wrapper;
 pub use metrics_reader::{Metric, MetricFormat, MetricsReader, MetricsReaderBuilder, Semantics};
-pub use varnish_macros::vmod;
-pub use vsc_wrapper::{Vsc, VscMetric};
+
+mod metrics_publisher;
+pub use metrics_publisher::{Vsc, VscMetric};
+
+pub use varnish_macros::{vmod, VscMetric};
 
 /// Run all VTC tests using `varnishtest` utility.
 ///
